+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Fix migration data permission [#531](https://github.com/rokwire/groups-building-block/issues/531)
+- Create posts migration api for switching the datasource to Social BB [#529](https://github.com/rokwire/groups-building-block/issues/529)
 
 ## [1.56.0] - 2024-01-23
 ### Added
-- Create migration api for Social BB [#527](https://github.com/rokwire/groups-building-block/issues/524)
+- Create migration api for Social BB [#527](https://github.com/rokwire/groups-building-block/issues/527)
 - Consolidate the information, and make it accessible with a single API call [#519](https://github.com/rokwire/groups-building-block/issues/519)
 - Implement GET /v2/groups API via http POST method [#524](https://github.com/rokwire/groups-building-block/issues/524)